refactor product listings page
